requirements.txt
boto3==1.10.28
botocore==1.13.28
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
Cython==0.29.14
docutils==0.15.2
future==0.18.2
idna==2.8
jmespath==0.9.4
joblib==0.14.0
kiwisolver==1.1.0
oauthlib==3.1.0
pyparsing==2.4.5
python-dateutil==2.8.0
python-twitter==3.5
requests==2.22.0
requests-oauthlib==1.3.0
s3transfer==0.2.1
six==1.13.0
urllib3==1.25.7
